Splitting up encoding base from encoding functions
